third_party/WebKit/LayoutTests/platform/mac/fast/forms/select/menulist-click.html

<!DOCTYPE html>
<body>
<script src="../../../../../resources/js-test.js"></script>
<script src="../../../../../fast/forms/resources/common.js"></script>
<select>
<option>option 1</option>
<option>option 2</option>
</select>
<script>

var clickEventCounter = 0;
var select = document.querySelector('select');
select.addEventListener('click', function() { clickEventCounter++; }, false);
hoverOverElement(select);
eventSender.mouseDown();
setTimeout(function() {
    select.blur();
    description('Check if mosuedown without mouseup on a SELECT element dispatches a click event.');
    shouldBe('clickEventCounter', '1');
    finishJSTest();
}, 1);
jsTestIsAsync = true;
</script>
</body>
