third_party/WebKit/LayoutTests/platform/mac/fast/forms/select/menulist-click-expected.txt

Check if mosuedown without mouseup on a SELECT element dispatches a click event.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS clickEventCounter is 1
PASS successfullyParsed is true

TEST COMPLETE
